test/bytestream: add explicit type casts
Fix a bunch of signed/unsigned comparison warnings in the bytestream
unit tests by adding explicit casts.  Some of the casts aren't strictly
necessary to fix the warnings, but they are added anyway for
consistency.

+++ src/lib/test/bytestream.cpp
@@ -7,10 +7,10 @@ TEST(ByteStream, U8) {
 	bs.write((uint8_t)57);
 	bs.write((uint8_t)137);
 	bs.seek(0);
-	ASSERT_EQ(57, bs.readU8());
+	ASSERT_EQ((uint8_t)57, bs.readU8());
 	uint8_t t;
 	bs.read(&t);
-	ASSERT_EQ(137, t);
+	ASSERT_EQ((uint8_t)137, t);
 }
 
 TEST(ByteStream, I8) {
@@ -18,10 +18,10 @@ TEST(ByteStream, I8) {
 	bs.write((int8_t)57);
 	bs.write((int8_t)-100);
 	bs.seek(0);
-	ASSERT_EQ(57, bs.readI8());
+	ASSERT_EQ((int8_t)57, bs.readI8());
 	int8_t t;
 	bs.read(&t);
-	ASSERT_EQ(-100, t);
+	ASSERT_EQ((int8_t)-100, t);
 }
 
 TEST(ByteStream, U16) {
@@ -29,10 +29,10 @@ TEST(ByteStream, U16) {
 	bs.write((uint16_t)0xa0b1, LE);
 	bs.write((uint16_t)0xc2d3, BE);
 	bs.seek(0);
-	ASSERT_EQ(0xb1a0, bs.readU16(BE));
+	ASSERT_EQ((uint16_t)0xb1a0, bs.readU16(BE));
 	uint16_t t;
 	bs.read(&t, LE);
-	ASSERT_EQ(0xd3c2, t);
+	ASSERT_EQ((uint16_t)0xd3c2, t);
 }
 
 TEST(ByteStream, I16) {
@@ -40,10 +40,10 @@ TEST(ByteStream, I16) {
 	bs.write((int16_t)-7340, LE);
 	bs.write((int16_t)0x5061, BE);
 	bs.seek(0);
-	ASSERT_EQ(21731, bs.readI16(BE));
+	ASSERT_EQ((int16_t)21731, bs.readI16(BE));
 	int16_t t;
 	bs.read(&t, LE);
-	ASSERT_EQ(0x6150, t);
+	ASSERT_EQ((int16_t)0x6150, t);
 }
 
 TEST(ByteStream, U32) {
@@ -51,10 +51,10 @@ TEST(ByteStream, U32) {
 	bs.write((uint32_t)0xa0b1c2d3, LE);
 	bs.write((uint32_t)0xe4f50617, BE);
 	bs.seek(0);
-	ASSERT_EQ(0xd3c2b1a0, bs.readU32(BE));
+	ASSERT_EQ((uint32_t)0xd3c2b1a0, bs.readU32(BE));
 	uint32_t t;
 	bs.read(&t, LE);
-	ASSERT_EQ(0x1706f5e4, t);
+	ASSERT_EQ((uint32_t)0x1706f5e4, t);
 }
 
 TEST(ByteStream, I32) {
@@ -62,10 +62,10 @@ TEST(ByteStream, I32) {
 	bs.write((int32_t)-123456, LE);
 	bs.write((int32_t)78910, BE);
 	bs.seek(0);
-	ASSERT_EQ(-1071776001, bs.readI32(BE));
+	ASSERT_EQ((int32_t)-1071776001, bs.readI32(BE));
 	int32_t t;
 	bs.read(&t, LE);
-	ASSERT_EQ(0x3e340100, t);
+	ASSERT_EQ((int32_t)0x3e340100, t);
 }
 
 TEST(ByteStream, U64) {
@@ -95,7 +95,7 @@ TEST(ByteStream, Float) {
 	bs.write(-1234.5678f, LE);
 	bs.write(1234.5f, LE);
 	bs.seek(0);
-	ASSERT_EQ(0x2b529ac4, bs.readU32(BE));
+	ASSERT_EQ((uint32_t)0x2b529ac4, bs.readU32(BE));
 	float t;
 	bs.read(&t, LE);
 	ASSERT_FLOAT_EQ(1234.5f, t);
@@ -121,13 +121,13 @@ TEST(ByteStream, writeVec) {
 	vec.push_back(0xd3);
 	bs.write(vec);
 	bs.seek(0);
-	ASSERT_EQ(0xa0b1c2d3, bs.readU32(BE));
+	ASSERT_EQ((uint32_t)0xa0b1c2d3, bs.readU32(BE));
 }
 
 TEST(ByteStream, readVec) {
 	narf::ByteStream bs;
 	bs.write((uint64_t)0x0102030405060708, LE);
-	bs.write(0x090a0b0c, LE);
+	bs.write((uint32_t)0x090a0b0c, LE);
 	bs.seek(0);
 	auto vec = bs.read(8);
 	ASSERT_EQ(0x08, vec[0]);
@@ -157,7 +157,7 @@ TEST(ByteStream, overran) {
 
 TEST(ByteStream, size) {
 	narf::ByteStream bs(7);
-	ASSERT_EQ(7, bs.size());
+	ASSERT_EQ((size_t)7, bs.size());
 }
 
 TEST(ByteStream, skip) {
@@ -173,7 +173,7 @@ TEST(ByteStream, tell) {
 	bs.write(std::string("abcd"));
 	bs.seek(0);
 	bs.skip(2);
-	ASSERT_EQ(2, bs.tell());
+	ASSERT_EQ((size_t)2, bs.tell());
 }
 
 TEST(ByteStream, bytesLeft) {
@@ -181,14 +181,14 @@ TEST(ByteStream, bytesLeft) {
 	bs.write(std::string("abcd"));
 	bs.seek(0);
 	bs.skip(1);
-	ASSERT_EQ(3, bs.bytesLeft());
+	ASSERT_EQ((size_t)3, bs.bytesLeft());
 }
 
 TEST(ByteStream, clear) {
 	narf::ByteStream bs;
 	bs.write(std::string("abcd"));
 	bs.clear();
-	ASSERT_EQ(0, bs.size());
+	ASSERT_EQ((size_t)0, bs.size());
 }
 
 TEST(ByteStream, setDefault) {
@@ -222,5 +222,5 @@ TEST(ByteStream, writeString) {
 	auto vec = bs.readString(narf::ByteStream::Type::U16, LE);
 	auto s = std::string(vec.begin(), vec.end());
 	ASSERT_EQ(std::string("abcd"), s);
-	ASSERT_EQ(12, bs.size());
+	ASSERT_EQ((size_t)12, bs.size());
 }
